Add sphinx_rtd_theme to enhance doc gate jobs
When it gets to submitting & reviewing patches for the project documentation,
it is very convenient to surf through the docs generated by Sphinx in a special
Gate job. The problem is that Sphinx uses in gates a theme which is different
from that used on ReadTheDocs. So the docs generated in gates look differently
from the docs that we have on ReadTheDocs, while the desired behaviour is to
have the same Sphinx theme both on ReadTheDocs and in gates.

There is a snippet that enables this theme for Sphinx locally (and so in gates):
http://paste.openstack.org/show/156901/

It is going to be used, for example, here:
https://review.openstack.org/120018

This code uses the sphinx_rtd_theme python package, so we need to include it
into global requirements.

Change-Id: I61dad599279b0dd94a70adcfd415f5a97701a5c4