Update patch set 4
Patch Set 4: Code-Review-1

Sean Dague and Matt discovered that this was an issue with SQL migrations in glance (specifically comments using /* */) and this can be fixed in master. Instead this patch should be applied to stable/{juno,icehouse} so that those versions continue to use a version of sqlalchemy-migrate that will not break those versions.

Patch-set: 4
Reviewer: Gerrit User 12000 <12000@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1