Update patch set 4
Patch Set 4: Code-Review+1

I'm a bit torn on this right now, we have the fix for glance on master now, but we need to pin sqlalchemy-migrate on the stable branches because the 006 sqlite scripts are in juno and icehouse and sqla-migrate versions aren't capped there.

And without this, someone could technically try running k-1 glance migrations with sqla-migrate 0.9.4 and it would blow up, so we *probably* still want this, right?

Plus if we merge this then we can use this to backport to the stable branches.

So I'd be leaning toward approving this.

Patch-set: 4
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1