Update patch set 4
Patch Set 4: Code-Review-1

We fixed glance on master, I think if someone is pinning to K1 then they can sort out the requirements themselves. Because if we block this then we are effectively rolling all the way back to sqla-migrate 0.9.1 again, and there are other fixes in here we want in the field.

Patch-set: 4
Label: Code-Review=-1