Update patch set 1
Patch Set 1:

I'm not sure I understand your concern Sean.  This wasn't a random incompatible release, this was a new major release which changed the major version from 2.x to 3.x.  The point of semantic versioning is to convey backwards compatibility changes.  It should be obvious that 2.x and 3.x are incompatible libraries.

Patch-set: 1