Update patch set 1
Patch Set 1:

My -1 remains because this still hasn't been communicated on the general ML. The only substantial barbican traffic in the last 6 months has been procedural (midcycle / core adds).

Changes which impact other projects *need* to be communicated on the ML. So we still need a postmortem email here explaining this incompatible change, why it was made, how the technical details of the new implementation will avoid this happening in the future.

Patch-set: 1