Update patch set 2
Patch Set 2: Code-Review-1

One of the perceived values of OpenStack projects all being in python is that you can move between projects and understand how different parts work.

At this point I'm actually a little concerned that we're running off in a bunch of different directions on how code is going to be written without any clear statement of direction.

This feels like this should be a TC discussion before we start changing the async programming model in projects. Can we get that scheduled on the agenda? I'd be more comfortable +2ing this if there was a statement of direction from the TC that all of OpenStack should convert to asyncio.

Patch-set: 2
Label: Code-Review=-1