Update patch set 2
Patch Set 2:

"At this point I'm actually a little concerned that we're running off in a bunch of different directions on how code is going to be written without any clear statement of direction."

asyncio was discussed multiple times on the openstack-dev mailing list, and in two blueprints (approved).

I wrote a global roadmap for my project "replace eventlet with asyncio in OpenStack":
http://aioeventlet.readthedocs.org/openstack.html

"This feels like this should be a TC discussion before we start changing the async programming model in projects. Can we get that scheduled on the agenda?"

I added "asyncio" to the agenda of the next meeting (January 20th, 20:00 UTC).

--

A previous version of my Oslo Messaging executor used *greenio* instead of aioeventlet. greenio and aioeventlet are almost the same thing.

The greenio dependency was only added in two days. You even approved it:
https://review.openstack.org/#/c/108637/

I don't understand why it's different for aioeventlet.

Patch-set: 2