Update patch set 2
Patch Set 2:

Please disregard my last comment about the review as I have abandoned that one. It is not possible to update the minimum django_auth_requirement version as that will conflict with master requirements which requires a minimum oslo.config>=1.6.0. The only way this can be solved is to either add the django-openstack-auth<1.1.8 constraint (this patch) or somehow release a 1.1.7 sub-version with the security fix.

Patch-set: 2