Update patch set 16
Patch Set 16: Code-Review-1

(7 comments)

I left a few comments about Oslo versions we should consider changing. I think the script you're using is pulling in the versions actually being tested, which is fine if they work, but if we're going to use stable branches for backports we should be consistent and use the last Juno release. (Some of the libs don't have the stable branches, yet, since for Juno we said we would create them as we needed them.)

Patch-set: 16
Label: Code-Review=-1