Adjust Oslo maintained library caps
For non-openstack maintained projects we are concerned about leaving
room for any future library versions, as we aren't sure they won't break
things.  But for oslo maintained projects this isn't an issue. The
OpenStack community is maintaining them and can make sure the releases
allowed here won't break things.

This is done as a follow up patch to pin everything, because tweaking
library versions is somewhat risky and we should land the pin everything
patch as soon as possible, and it is currently working.

Change-Id: Ic1b9733f0602228bf79620411d495fad9473203f