Update patch set 1
Patch Set 1: Code-Review-1

so... on 2nd thought, this is incredibly silly. I can't imagine any coinstalled projects ever going to depend on python-openstackclient in a way that the versions need to be synchronized.

As far as I can tell devstack is the only thing that cares on version, because of cli arg usage. Devstack isn't synced in global requirements because it's not python. We should actually remove this instead and just leave the hardcoding in the right place in devstack, because it's completely silly to move approving this change in devstack to require approval in g-r first.

Patch-set: 1
Label: Code-Review=-1