Update patch set 1
Patch Set 1:

So this isn't going to work right now.

devstack on stable/icehouse installs python-openstackclient <= 1.0.1 and 1.0.1 has uncapped python-novaclient so it's pulling in novaclient 2.21 which has a min version on keystoneclient that conflicts with the capped version of keystoneclient in stable/icehouse global-requirements, and then kaboom.

The solution, I'm told, is we need to run python-openstackclient and other things in a virtual environment like we're doing with tempest(-lib?) now, and that work starts here:

https://review.openstack.org/#/c/157535/

Patch-set: 1