Update patch set 1
Patch Set 1:

So I don't see a wheel for routes on pypi: https://pypi.python.org/pypi/Routes/2.1

I also haven't  seen any of the failures in dsvm jobs related to this anywhere in the gate. Routes 2.1 was released on Jan. 17, so if this were a problem I would have expected us to have already fix it. Is there a bug filed that has this info?

It's also worth noting if this is really needed, you'd probably have to backport this to juno as well.

Patch-set: 1
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0