Restore minimum requirement version for kombu
Commit 499db6b64071c2afa16390ad2bc024e6a96db4ff capped anything that
didn't have an upper bound for stable/juno using what was currently
tested in the gate from pip-freeze output.

However, it pinned kombu at 3.0.7 when it was >= 2.5.0 before which
breaks anyone that's already shipped a Juno-based distribution with the
level of packages required in the last release (2014.2.2).

This change restores the lower bound on kombu and makes 3.0.7 the upper
bound but not the only required version.

Going through the other == pins in commit 499db6b64 the only other one
is suds but it pins on what was already the minimum required version,
0.4, so that was OK.

Change-Id: I8bb1c2a752156271a88b832863b1b6fc26e61e7f