Adding weasyprint to download Horizon reports in PDF format
WeasyPrint is a pure-python library for PDF generation, based on CSS
and WEB standards. It supports Python 3 and comes with a full test
suite.

This feature is useful for capturing the updated information of
instances, volumes, flavors, etc at the moment in a PDF report,
which can later be used for sharing/mailing information.

This patch is required for https://review.openstack.org/142676

Change-Id: I311eb4b09f1c09ca3f7d2a5eac6f9199bdca184a
Partially-Implements: blueprint horizon-reports-in-different-formats