Update patch set 2
Patch Set 2: Code-Review+2

This looks like it may be the same problem that prompted 0.8.7 to be blacklisted.  I'm okay with this, but would be even more okay with something stronger so we aren't playing "whack-a-mole".

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2