Update patch set 5
Patch Set 5:

This change raises the minimum versions for the other libraries to reflect what we're actually testing with in our gate, and to ensure that all projects that are relying on Oslo features introduced in Kilo accurately reflect those dependencies.

We've already capped the requirements on stable branches, and some of the minimums in master were already higher than those caps. That means any colocated services using those libraries need to be updated together, so this change expands that restriction but doesn't introduce it.

Patch-set: 5
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0