Update patch set 5
Patch Set 5: Code-Review-1

We don't, as a general rule, cap the upper bounds of libraries in master. We *want* to be testing with new releases, to find issues before we release code into the wild.

We do make exceptions in cases where upstream packages are known to be problematic. Is that the case here?

Patch-set: 5
Label: Code-Review=-1