Update patch set 5
Patch Set 5:

@Peter - I understood what you were doing, although I didn't realize you (or we?) managed the library. My point was that we specifically have a policy against capping libraries unnecessarily, and especially if we have control over the code.

Having things be broken once in a while is just part of the development cycle. In the case of upstream libraries, it helps us identify incompatibilities in the way we're using code or persistently problematic upstream developers. I don't think either of those describes this library so I don't think we need the cap.

Patch-set: 5