Update patch set 1
Patch Set 1:

As Brant said, this is something we are already relying on. It also shows that we have a gap in our testing coverage. Ideally, we should be running a set of tests with the minimum versions in the requirements.txt. At the very least for a unit test run.

Patch-set: 1