Update patch set 1
Patch Set 1:

I don't think we should gate on anything that is allowed to change underneath us.  So, to say we shouldn't gate on pylint because it can change is nonsense.  Anything that we allow to change underneath us will probably break us eventually, contract or none.  It frightens me that anything is allowed to change below us and break tests that previously passed.  We've lost a lot of gate time to things changing from underneath.

Using pylint in an advisory mode isn't very valuable.  It will likely be broken more often than not.  So, if we don't care to gate on it then I'd rather just remove it.  It seems that it has added some amount of value to the project's code quality, so I guess I lean towards keeping it.

What we cannot do is leave it like it is.  So, either remove pylint entirely or pin it.  Those are the only two viable options.

Patch-set: 1