Update patch set 1
Patch Set 1: Code-Review-1

We pin hacking, but we have enough control over it to ensure that new versions are not introduced before projects can safely use them. We have seen several cases where new releases of pylint introduce checks that are not backwards compatible and cause existing jobs to fail where they were passing before. As Sean points out, this makes it a bad practice to rely on using it in a strict-voting job. We need to consider that other teams may want newer versions of pylint, and so they wouldn't want a pin or a cap put in place here. That's why most or all of the other projects using pylint have the job set to non-voting.

As far as rules go, there is no single hard and fast rule for deciding what should be pinned. As I said, hacking is capped but we have control over that code and can manage updates smoothly. All of the other projects with caps appear to be either runtime dependencies with known issues, build tools with known issues used where the job has to run successfully (sphinx), or Oslo libraries being capped as we prepare to create the kilo stable branches of the apps. As a style-checking tool, pylint is not like any of those things.

Patch-set: 1
Label: Code-Review=-1