Update patch set 1
Patch Set 1:

All projects that depend on anything in this file must be able to get along with whatever the version du jour is.  The question here is whether we control that version du jour ourselves or whether we let someone else do it.  So, the version is effectively pinned, we just don't have control over what the pin is and when it changes.  (and somehow this is better?)

Right now, we can move a version forward because we relinquish the control of the pin to an outside third party.  Whenever they change the pin, it changes for all projects.  Then, when moving the pin forward causes a problem, we react immediately because everything is broken and no one can get anything done!  This sort of thing might've worked well when the project was smaller and younger but I don't think it is the right way to go forward.

We need a way to bump versions in a global way similarly to what we have now except that we put control over the version pin in our hands and out of third parties hands.  That way, bumps in versions can be tested across all projects in a quarantined environment that doesn't halt everything when it breaks.  For the requirements with good contracts that aren't expected to break anything, this will still be routine.  Just bump the version, test, and merge.  A bot could do this.  For requirements which break things -- like pylint or whatever else -- we quarantined the version bump to give time to work out the issues without killing very valuable time in our gate.

Patch-set: 1