Update patch set 1
Patch Set 1:

If we want to move to a new version of pylint, we just bump pin here and check that it does not generate issues in projects when OpenStack bot syncs the new pin to their respective trees. If there are issues, projects are free to fix them and then bump. So we're not breaking anyone with the pin.

Patch-set: 1