Update patch set 1
Patch Set 1:

@Sean,

This will make the minimum version of cryptography requied by keystone not work with the Fernet token backend. This, gap was strictly because we did not test this with the minimum version.

I would say the minimum making an entire feature not work is a release critical bug.

Patch-set: 1