Keep pyOpenSSL < 0.14
The 0.14 version of pyOpenSSL has been rewritten to use "cryptography"
(the package) and its new method of binding to the openssl library
consumes about an order of magnitude more memory (2mb -> 20mb).

This is enough to cause OpenStack CI virtual machines to hit OOM
conditions during testing.  For more investigation into this see [1];
note that the CI ubuntu images currently come with the packaged 0.13
version pre-installed which is why only CentOS started seeing this
problem.  Ubuntu is, however, affected if it uses the 0.14 version.

An issue has been raised with pycparser which is currently under
investigation [2].

[1] https://etherpad.openstack.org/p/oom-in-rax-centos7-CI-job
[2] https://github.com/eliben/pycparser/issues/72

Change-Id: Ifb708ba1c6d1d53c54275753489b91fbe9503739