Update patch set 1
Patch Set 1:

I think I need someone to explain some of this :-)

For kilo there'll be a kilo/stable branch for glance_store - right? And I'm guessing that releases off that branch will be 0.4.1,0.4.2,0.4.3...?

If so, typically I'd expect that the 0.4.x releases would be only targeted at the stable/kilo glance.

Are we assuming that any 0.4.x release must be appropriate for both stable/juno and stable/kilo?

Thanks

Patch-set: 1