Update patch set 3
Patch Set 3: Code-Review-1

This seems like a huge jump in required versions for a "stable" branch.  As a deployer that is following the stable branches this would have an impact on anyone looking to deploy juno after the great pinning change which happened Jan 15 - "https://github.com/openstack/requirements/commit/499db6b64071c2afa16390ad2bc024e6a96db4ff"

Pinning to glance_store >= 0.3.0 also brings in a new dependency that is not found within the current Juno global requirements.

IE: enum34 - this is not in global requirements for Juno at the head of the stable/juno branch.

If anything is added here I'd like to see this be a requirement allowing all versions that have been dep'd on within the Juno cycle.

IE:  glance_store>=0.1.1,<0.4.0

However, I have to disagree with this change because it adds a dependency to the "stable/juno" branch which is not accounted for in global requirements and is a huge requirement jump. IMO If `glance_store` is going to be providing stable branches in the kilo cycle then this change should really be focused on Kilo only and have no impact on Juno.

Patch-set: 3
Reviewer: Gerrit User 7353 <7353@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1