Add os-testr to global requirements list
This commit adds os-testr to global requirements since it will be
the default runner of testr for most projects' tests. Also the direct
testrepository (and possibly subunit) requirement will be removed from
projects in favor of letting os-testr dictate that requirement. So it
is also added to the projects list.

Change-Id: Ia399c65e62c2306c9a2d082c6219fcb030b5946b