Update patch set 2
Patch Set 2:

i believe i have a workaround for ceilometer. that said, some of the deprecated functionality, specifically 'ensure_index', behaves unpredictably with the new version as it caching workarounds to support backward compat[1]. we use 'ensure_index' quite a bit in ceilometer so we need to update all the references for when the pymongo decides to remove the method completely.

[1] https://github.com/mongodb/mongo-python-driver/commit/f93c73aa7391c2037758e2990b4930a1d4e935d5

Patch-set: 2