Update patch set 1
Patch Set 1:

See the message from Akihiro here [1], if we've done something horribly wrong I apologize, but we wanted to ensure that 2.3.x was a branch for Juno, 2.4.x was a branch for Kilo, and we'll use 2.5.x for Liberty. This was from an earlier break we made in February [2] when we released 2.3.11, this strategy seemed like the right approach.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-April/061089.html
[2] http://lists.openstack.org/pipermail/openstack-dev/2015-February/057384.html

Patch-set: 1