Update patch set 1
Patch Set 1:

We don't usually bump minimums just for bug fixes. This is a security-related fix, and so we want deployers to have it, but the existing requirements entry should allow it already. Do you think we need to make an exception to the general bug-fix policy in this case?

Patch-set: 1