Update patch set 20
Patch Set 20:

did anyone check if the some of the requirements actually make sense?  for instance, requests has a exclude that's outside of the range of the allowed.

+requests>=1.2.1,!=2.4.0,<=2.2.1

Also, was there discussion with packagers about this change?  It's making openstack very difficult to package for and I was already considering dropping it.  It seems that openstack can't figure out what dependencies it actually wants to use.

The capping is likely fine from a packager prospective, at least for the distro I'm packaging for (we have some flexibility for versions that get installed).  What's thrown me (and likely others) for a loop is that we had no input into this lockdown.  As long as the locking is done at the outset of the first release tag (juno/kilo/etc) it makes it easier to support.  This change in the middle is nearly impossible and made me come VERY close to removing all openstack services.

Patch-set: 20
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0