Mark documentation repositories as soft
Since not all requirements are wanted in the global-requirements, the
update.py script from requirements has a soft-update parameter to
ignore requirements that are not in global-requirements.txt.

Once the dependend change for project-config merges, we can mark
projects as soft.

Mark all documentation projects as soft and remove dependencies from
global-requirments that are only used by these projects and not anywhere
else.

Change-Id: I807b061cf697dc79b26ffdfbe023c32600ddd346
Depends-On: I57a4b9c7471458ae7ba1c98db463e41afe7dc6b2