Update patch set 1
Patch Set 1:

With this you'll be able to propose changes to stable/juno, but you're still not going to be able to do a release that includes those changes. The next fix release after 1.1.1 is 1.1.2, and with the cap a 1.1.2 you won't be able to propose any changes to it just like you weren't able to propose changes for 1.1.1.

Patch-set: 1
Reviewer: Gerrit User 6486 <6486@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0