Update patch set 20
Patch Set 20:

For some more context see: http://lists.openstack.org/pipermail/openstack-dev/2015-February/057068.html

In short, this was a stop gap measure and we will be capping stable/kilo right away not in the middle of the cycle.


Also, what is the process of updating the freezes?  -we should update packages individually.  As you have already noticed https://review.openstack.org/#/c/161047/ is the more long term solution.

Is there a particular list where we can discuss packaging? -- Maybe just on the dev ML?

On thing that was nice about only setting the caps on known failures is that it made it MUCH easier to package for. As long as we don't have stable branches breaking underneath us and causing master to be blocked due to grenade that is great. But the previous process didn't work

Patch-set: 20