Update patch set 1
Patch Set 1:

Hi Joe,

That change was for glanceclient, and if you look into Nikhil's comment you can see they plan on re-enabling it.

On glanceclient they were only running Juno gate (and it was failing), but in cinderclient we are running Icehouse (which is running fine) and Juno (which is failing) so I thought it would be better to allow the gate to run.

Patch-set: 1