Update patch set 1
Patch Set 1:

I'm clearly missing something, because I thought we wanted to ensure backwards compatibility in cinderclient and that's why we were running gates for juno and icehouse.
And even if we create a stable/juno branch in cinderclient, wouldn't we still get something like version 1.1.1.post100 from the library, as we are getting now, which would not be allowed by the dependency capping?

Patch-set: 1