Bump kombu version and add amqp version
oslo.messaging now implements rabbitmq heartbeat.

py-amqp (the underlying low library to connect to rabbitmq) has a
working support of heartbeat since 1.4.0. So we need at least
that version.

Given that Trusty has 3.0.7 of kombu and the authors have indicated
that amqp>=1.3.3,<2.0 would work with that version, we can use
amqp 1.4.0 with kombu 3.0.7.

So this change raises the kombu requirement to '>=3.0.7' and
amqp to '>=1.4.0'

Change-Id: I2b111af06e27caf1436cbda2de5be547c7d8790c
Closes-bug: #1436769