Remove doc repositories from projects.txt
The documentation repositories maintained by the OpenStack
Documentation team are currently all in projects.txt. They
are not installed together OpenStack server projects.

While the syncing of requirements is nice, it's not essential that
Documentation repo are part of requirements. The Documentation
repos have additional unique dependencies that are not needed by
other projects.

Still, the Documentation team likes to ensure that their requirements
are not conflicting with global requirements to allow editors and
developers reuse environments.

As discussed with Documentation team, remove these repositories
now (see thread starting at
http://lists.openstack.org/pipermail/openstack-docs/2015-April/006362.html).

Change-Id: I807b061cf697dc79b26ffdfbe023c32600ddd346