Update patch set 1
Patch Set 1:

python-cinderclient is used in two ways:

* A python library
* A CLI

For the CLI it should work with all supported versions of cinder. Previously when the python library part of cinderclient was used on stable branches in a uncapped fashion we had to make sure ensure backwards compatibility. But making old versions of code support software that wasn't released at the time of cutting the stable release was a huge pain. So now that we cap stable branch's dependencies we do not need backwards compatibility. We need to make sure cinderclient's stable/juno branch is compatible with juno and the same for kilo. But we don't need master to work with stable Juno.

And yes this change is needed if you have a stable/juno branch of cinderclient.

Patch-set: 1