Update patch set 3
Patch Set 3: Code-Review+1

I haven't been able to figure out why some environments end up with 2.2.1 installed and some get 2.6; I *think* this is down to Pip ignoring dependences the second (and subsequent) times it sees the same package listed. IE, I think that some environments end up with 2.2.1

We might be able to work around this by changing the way Magnum is installed to make sure that it always ends up with a version of requests that works for it; but even if we do that, it seems that version 2.2.1 doesn't actually work for us (in all cases - it seems to be fine for some cases) any more. On that basis I think that bumping the listed minimum is justified.

Patch-set: 3
Reviewer: Gerrit User 9453 <9453@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1