Update patch set 1
Patch Set 1: Code-Review-2

If older versions of glance client work with HTTP but not HTTPS, we can consider this an important bug fix, but still a bug fix. As such, we don't need to hold up the release to manage a minimum requirements change. We should *definitely* write a release note about recommending using a newer version, but for configurations where HTTPS is not needed between nova and glance the versions we've been testing with should be OK.

Patch-set: 1
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2