Update patch set 3
Patch Set 3:

Sean - this was the suggestion of a chat we had with some of the fine folks in infra a few months back, who correctly pointed out that linters are known to break gates when new features are added.

The consensus was that a pinned version which can optionally move up once per OpenStack release was the best approach.

Patch-set: 3
Label: Verified=0