Update patch set 1
Patch Set 1:

For the record, ttx, sileht and I decided on IRC to release oslo.messaging with the feature turned off and make a note in the requirements explaining that when the feature is enabled the version of kombu needs to be updated. This avoids needing to update the minimum requirements of all of the projects using oslo.messaging for kilo, while still allowing deployers who want the feature to have a path to configure it.

Patch-set: 1
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0