Update patch set 4
Patch Set 4:

This version contains the ability to load raw configuration values into the config fixture, which was a special request by dhellmann during an update to oslo.middleware. So to answer your question: no, it's not strictly necessary, but I can't update the CORS middleware tests in oslo.middleware without it.

Patch-set: 4