Update patch set 1
Patch Set 1:

Switching to actually using pylint from requirements/global-requirements would be fine too - my point was that _this_ change in isolation is not sufficient to achieve the stated goal.  I haven't seen a corresponding change to move neutron to using pylint from requirements mentioned here, so apologies if it already exists.


The downside (and I think it's a big one!) of moving pylint to global-requirements is that then every project has to agree to use the same version of pylint.  There's no actual technical reason that needs to happen (because pylint is only a dev-time requirement and all dev happens in separate tox venvs), so we're just adding future procedural costs from coordinating future pylint upgrades.

Because of the above, I'm -1 on this global-requirements approach, and suggesting we instead just pin pylint directly in neutron tox.ini.

Patch-set: 1