Add mysqlclient dependency
mysqlclient is a MySQL driver supporting Python 2 and Python 3.

mysqlclient is a fork of MySQL-python. It was created by INADA Naoki,
who is also the author of PyMySQL (a pure Python driver for MySQL).
mysqlclient adds Python 3 support, some new features and various
bugfixes.

mysqlclient and MySQL-python are exclusive, they cannot be used at the
same time (with the same Python version), since both packages provide
the same Python module ("MySQLdb"). An application can use MySQL-python
on Python 2 and mysqlclient on Python 3, or mysqlclient on Python 2 and
Python 3.

mysqlclient is distributed under the same license than MySQL-Python: GNU
GPL.

Change-Id: I109f3726fff27de2b61bc3423615207c12bb0af2