Update patch set 2
Patch Set 2: Code-Review-1

Not thrilled with this. The outcome from the mailing list was that using it for python3 installs sounded like a good idea, but this does nothing to do that. Nor do we have a great answer for that until markerlib comes out. I'd like to see that sorted out before we land this and accidentally migrate something in the gate that we were not expecting to migrate.

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1