Update patch set 2
Patch Set 2: Code-Review+1

Monty, the issue is that MySQL-python is unmaintained, while mysqlclient seems to be. I don't agree that having MySQL-python for Py2 and mysqlclient for Py3 is "the outcome" of the mainling list discussion, this is only the opinion of Mike Bayer, and I don't agree. I'm on the side of Victor here.

Patch-set: 2
Reviewer: Gerrit User 6476 <6476@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1