Update patch set 2
Patch Set 2: Code-Review-1

I'm not sure what it means to have both mysqlclient *and* mysql-python in the requirements file at the same time, since they use the same module namespace (correct me if I'm wrong on that).

Also, this would mean that distributions would need to *uninstall* their current mysql-python packages to make room for this one (again, I'm not a distro expert, but that's what it seems like, correct me if I'm wrong.  i'll try to ask around downstream here to see what the implications are).

AFAICT mysql-python is still maintained as far as critical bugfixes.  Other than "does not support eventlet", which mysqlclient has just as much, I'm not familiar with any bugs in it that are impacting openstack applications right now (again, I'd love information on this).

As Python 3 is where things are headed, as installations migrate to Python 3 which is still more bleeding edge, they can also upgrade to the less proven MySQL client as well. 

if *anything* were to change in Python2, it should be that we go to PyMySQL, or drop eventlet.   Using another MySQL driver that is just as broken with our concurrency model as the current one, *plus* introducing the packaging issue and whatever instabilities are present by the fact that this is a much more actively changing driver, seems crazy.

Patch-set: 2
Label: Code-Review=-1