Update patch set 2
Patch Set 2:

> Which critical bugfixes were fixed recently?

what critical bugfixes *are* there?  

I've proposed to mysqlclient that they change the import name of their package: https://github.com/PyMySQL/mysqlclient-python/issues/44  Then the issues here go away entirely.

It should not be impossible to run both DBAPIs at the same time, and using the same name here is bad form.   The psycopg2cffi driver for pypy is the only psycopg2 that you can use on pypy, yet they still went with a different import name.

As far as linux distros, I had missed Thomas' reply to me.   But I'm not sure how we're comfortable with the idea that one package would silently replace the other, and not even deterministically!  What if I install the mysqlclient package before the mysql-python one?  Now I have the *old* version in place.  How is that to be accounted for?

Basically if mysqlclient becomes a good python citizen and picks their own module name, go nuts.  We all saw what a nightmare it was when distribute (necessarily) had to use the same module name as setuptools.

Patch-set: 2