Update patch set 2
Patch Set 2: Code-Review+1

As noted in the commit message, this is more of a refactoring than a completely new project.  We've known for a while that there was code in oslo.concurrency that could be generally useful, but was awkward to consume outside the OpenStack ecosystem.  Happy to see progress being made on splitting that out into its own thing.

Maybe somewhat unrelated question, but do you plan to bring this in to the Oslo program at some point?  It would be good to have it included in our CI.

Patch-set: 2
Label: Code-Review=+1