Update patch set 1
Patch Set 1: Code-Review+1

Yes please. We (ibm) ran into some major uglies with packaging rpms because of the min version on pbr 0.6 so we build against 0.6 and run against 0.6 as an rpm, but in our test nodes we don't package tempest so we clone it from github and then pip install tempest's requirements, which pulls in pbr 0.11.  pbr 0.11 is enforcing versions the way that pbr 0.6 is not (we have a 2015.1.ibm.rc1 tag which is allowed as a version with pbr 0.6 but not 0.11), and things explode in not so obvious ways.

Patch-set: 1
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1