Uncap pytz and add a note on ordering between Babel and pycadf
Change Id18916953f6e1bae3688216eaf809c2086b4af13 is failing the
stable/juno runs because Babel 1.3 is installed first which installs the
latest pytz (pytz is uncapped in Babel 1.3), but pytz is capped in
pycadf 0.6.2 (latest that will run under stable/juno g-r), so we hit a
ContextualVersionConflict.

Add a note in global-requirements.txt warning users about the order.

This is not a problem on master or stable/kilo because pytz is uncapped
there.

We uncap pytz here because (1) stable/icehouse has a completely opposite
supported version range for pytz in g-r which makes grenade sad and (2)
pytz does not have dependencies of it's own, it's a fairly simple
project so we don't have a transitive dependency risk here so uncapping
should be safe.

Change-Id: I3ca651a4377d61be3fed572355326bc180ad7eec