Update patch set 1
Patch Set 1:

apevec: We've never maintained a stable branch for python-heatclient, I don't even know why that branch exists.

To my knowledge, none of the other clients observe a stable branch model either, so AFAIK if you want bug fixes you have to upgrade to a new version, which is also the reason we're very careful to maintain backwards compatibility in the clients.

Patch-set: 1
Label: Verified=0